requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
rank_bm25==0.2.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy<2.0.0
scikit-learn==1.4.1.post1